requirements.txt
# DWMB core
pydantic>=2.0
python-dotenv>=1.0

# Optional: Supabase (for storing instances/runs)
supabase>=2.0

# Optional: PPO+LSTM baseline
torch>=2.0

# Optional: tests
pytest>=7.0
